Switch to iframe, add SVG, buttons and light sensor.
An iframe seems the most viable way to embed this and matches the
approach to simulation in MakeCode. I've kept the terminal on the
outside. Just a PoC at this point.

The display now scrolls on a micro:bit SVG. For the moment it's loaded
via an object tag but more could move to JS over time.

For buttons, I've kept the C hal code that maintains separate state for
presses and was pressed but it could be approached differently.

Sensor UI might not belong inside the iframe, might make more sense to
define the sensor types and an API, but I've put one in for now to
experiment.